fix(server): removed the /app and its contents, /app already existed in the /src dir
